Prevents leaking state across plugin instances
State information was being stored on the element, which could be the
target of multiple instances of hoverIntent. Because there was no way of
knowing which instance the data came from, it was possible for instances
to step on each others' toes and overwrite this stare information,
thereby preventing execution or accidentally triggering incorrect behavior.

This fixes the problem by assigning a unique ID to each hoverIntent
instance and then using this ID to store per-instance state information
on target elements, cleaning up once hovered out.

Some refactoring was also done to clean up function scope and improve
memory usage, and many comments were added and reworded for clarification.

Fixes briancherne/jquery-hoverIntent#35.