copy current mouse coords to `over` handler event
This overwrites the pageX and pageX properties of the mouse event passed
to the `over` handler with the coordinates obtained from the most recent
tracked mousemove. This is usually more useful than the possibly-stale
coordinates of the pointer at "mouseenter" time.

Thisi change may break compatibility with any handler that expects the
original coordinates, though these are still available on the
event.originalEvent.

Closes briancherne/jquery-hoverIntent#11.